fixed id bug

+++ Game-idea/gameinfo.hpp
@@ -12,7 +12,7 @@ struct LobbyInfo {
 	sf::Uint16 stone = 0;
 
 	Player player;
-	std::map<sf::Uint16, Player> otherPlayers;
+	std::map<sf::Uint32, Player> otherPlayers;
 };
 
 //stored on the server
